docs(server): Replace send_file example.
